Split JudgementOutcome into its own file
