+++ Assets/Scripts/Judges/Judge.cs
@@ -49,17 +49,3 @@ public sealed class Judge
         return new JudgementOutcome(judgement, intensity, dt <= miss);
     }
 }
-
-public readonly struct JudgementOutcome
-{
-    public JudgementOutcome(Judgement judgement, float intensity, bool shouldConsumeNote)
-    {
-        Judgement = judgement;
-        Intensity = intensity;
-        ShouldConsumeNote = shouldConsumeNote;
-    }
-
-    public Judgement Judgement { get; }
-    public float Intensity { get; }
-    public bool ShouldConsumeNote { get; }
-}
